@j-bac 2024-03-11
thanks Jesse. Really nice speedup!