@JesseCresswell 2024-03-11
Those were all the changes I planned to make. It would be great if you could make a new release as I'm actively using the repo for research.