@j-bac 2024-03-11
Glad to hear it's useful. Just created a new pypi release